requirements.txt
ccxt>=4.1.22
python-dotenv>=1.0.0
google-generativeai>=0.3.0
pybit>=5.6.0
sqlalchemy>=2.0.0
requests>=2.31.0
pytrends>=4.9.0
textblob>=0.17.1
flask>=3.0.0
pyngrok>=7.0.0
yfinance>=0.2.35
numpy>=1.24.0
pandas>=2.1.0
fredapi>=0.5.0
pytz>=2023.3
ta-lib>=0.4.0
plotly>=5.18.0